Don't let level excursions affect env.old_player_pos
Independent of the related bug fixed in 012bc58e2 (which had the same
practical impact), this let you use a level excursion to force a monster
to forget you, because it saved the wrong position after the level
excursion. A bit hard to use effectively, but would have an effect on
any out-of-los monsters when stairdancing on a single stairs, for
example.