Increase size of zig permastore the easy way
With zig ghost floors, the game often needs to pull a lot of ghosts at
once. This commit increases the backup permastore size to the max (127)
by simply copying existing permastore ghosts from pan, tomb, and depths.
This is in lieu of a more involved ghost update which will take longer.

I chose to include some depths bones (about 15) to give a tiny bit more
level spread. It's still very high but this gives a small chance to have
an xl 15-25 ghost from time to time, if the floor is built drawing on
the permastore. (The average death xl for recent is 26.64 so it's
possible I've lowered this too much, but the average will still be quite
high.)