Don't let Xom add to weight 0 innate mutations (Lightli)
Other code paths check weight differently than this but Xom uses the
weight calculations in _accept_mutation, and they provide a small chance
of adding on to any innate mutations that are weight 0. One could
imagine that this might have once been intended (because Xom) but it's
not written in such a way that makes me think this is likely; I'm also
pretty sure Xom can't remove such mutations.