Disable cursor validity checks on windows
This obviously isn't ideal, but I can't figure this out with the limited
tools / speed I have in my VM. Probably what is going on is that one of
the lower-level windows-specific drawing calls is leaving the cursor in
an odd state before a viewwindow call.

This change should be basically safe in that the problem isn't in the
higher level code, which at this point is quite heavily exercised on
linux/mac.