Lower the spell fail polynomial cap
This was introduced in cef5c758032e to deal with potential overflows in
low strength / heavy armour scenarios. However, this commit didn't take
into account the fact that `chance2` is signed, so while I'm not sure if
an overflow was possible in practical scenarios, it was possible in
principle. (CanOfWorms and I discussed this at the time but it looks
like neither of us ever made the commit.) This commit lowers this cap
substantially so that this polynomial is always dealing with pretty sane
values.

Resolves #1420