Don't force screen update in print_stats
This commit splits screen updating out of stat printing, requiring the
caller to update the screen explicitly. This is analogous to viewwindow
for local tiles, where viewwindow does not render to the screen, but
instead loads a dungeon viewbuffer into the dungeon widget, which then
renders it on the next redraw.

Printing to the screen and actually showing those changes ('swapping the
buffers') are completely different actions that should not be entangled.
Printing will be done by widgets, but updating should be done only by
the ui framework. Even for legacy bits of crawl code that do a lot of
manual view updating, and the hacks that join the new code and the
legacy code, it's much better to be explicit since this allows doing
multiple paints and then updating the entire screen in one fell swoop;
this is critical for eliminating rendering issues and ttyrec bloat.

Next: viewwindow, redraw_screen