Don't update screen when exiting last UI layout
The point of this redraw_screen call was to restore the contents of the
screen to a known state before returning control to the main game loop
and its associated rendering code. Since it's going to update the screen
anyway, there's no point in us doing it as well, and it's better not to,
as it will reduce flicker.