Don't log initfile errors on webtiles (+refactoring)
The options file is read in two passes: one before the whole io system
is initialized, one after. For offline purposes, any errors before the
io system are initialized are echoed to stderr, in case the game never
succeeds at io initialization. In addition, a bunch of specific errors
for certain options were explicitly sent to stderr for reasons that
aren't clear to me (these fprintf lines go back to linley as far as I
can tell). Anything sent to stderr gets logged by webtiles and so these
options errors both clutter up the log and unnecessarily alarm server
admins, where the only meaningful signal in them should be to the player
("your rc file needs updating"). This has particular been an issue
because non-updated but widely-used HDA rcs set a color for the pray
channel (removed a while ago) and the food channel (removed recently).
The contentful part of this commit is to (a) standardize these initfile
errors to use the same reporting mechanism as everything else (and so
get rcfile line #s), and (b) suppress all echoing to stderr for anything
that happens on the first initfile pass for webtiles builds, which
should lead to no options errors getting logged. Players will still see
the errors when starting up their game.

In addition, I did a bunch of refactoring and moved a tiny bit more
messaging-related code into the existing, but barely-used, `msg`
namespace. The most public-facing api change here is that `no_messages`
is renamed to `msg::suppress`.