New option: travel_one_unsafe_move
Setting this option to true causes auto-travel/auto-move to skip safety
checks for the first move. This is intended as an accessibility option,
allowing easier expression of things like "take one step to the nearest
stair" via the user interface for users who would like to or need to
rely more on automated pathfinding.