Split const and non-const parts of describe_item
describe_item gets called in two contexts: describing an item that can't
be interacted with at present (in view, in a shop, an acquirement offer)
and describing an item in inventory (which offers interaction commands).

The former use case is completely const, the latter is necessarily
non-const because the available actions might modify the item in some
way.

The combination of these use-cases led to several unnecessary
const_casts in the codebase. This commit splits the two use cases into
two separate functions and updates all call sites that only want the
former behavior. It also explains a const_cast used in shopping and
acquirement, and removes an unneeded unwind_var in those settings.