Prevent Wu Jian martial attacks while rampaging or digging
"Fancy" moves have strange interactions with Wu Jian martial moves; when
it was just Dig (works) and Hop (doesn't) the inconsistency wasn't that
noticable. With the addition of Palentonga charge and boots of
rampaging, neither of which work well mechanically with Wu Jian, the
consistency issue becomes more strained.

This commit cleans up the rampaging interaction with Wu to avoid a host
of complications, and tweaks martial moves to only be compatible with
"normal movement" - disabling Wu processing while digging. Wu Jian's
description is accordingly updated.