Don't crash on quiver drawing during game setup
This is not very easy to trigger, but as reported in #1503 an accidental
change in spell starts allowed it to be possible to trigger this bug
with an appropriate force_more_message. Basically, if the hud rendering
code is called before the inventory is fully setup, there will be a quivered
item that is not yet linked, and this call would crash. Instead, just
silently return -- it's probably not going to be what is intended to
draw the hud in this state, but it's always happening by side effect so
shouldn't crash.