Don't unconditionally draw noise/equip bar
This is mainly aimed at the fact that this bar, unlike everything else
on the hud, unconditionally rendered, so when starting up the game it
would show up on its own at odd times. This commit changes the noise bar
so that it doesn't render during startup, but renders unconditionally
the rest of the time -- smarter logic is no doubt possible, but I was
somewhat defeated by silence. The equip bar only redraws as needed by
gear or weapon changes.

This also removes an unused return value in output.cc and some support
code.