Add github issue tracker links in a few key places
It doesn't seem like mantis is going to be generally usable any time
soon, so this makes sure that github issue tracker is linked both in the
manual and in crash messages. I didn't go through all the mentions of
mantis in the dev-oriented docs. This isn't intended to deprecate
mantis, because we do use it and accept bug reports there for anyone who
has an account (and there's a *ton* of history as well).

Resolves #1506.