Add glyph mode customization options for webtiles
This commit adds two options `glyph_mode_font` and
`glyph_mode_font_size` that allow customization of the font and its size
used to render the dungeon view in the webtiles glyph mode. If set to
monospace (default) or the font is not available, the browser's default
monospaced font will be used.

While implementing these options, this commit also follows the glyph
metrics exactly as prescribed by the font. With the change, the dungeon
view of the console version and the webtiles glyph mode should be more
similar and the box-drawing characters will no longer appears to be
unconnected, provided that the monospaced font used by the browser is
properly authored in the first place.

[ Committer's note: Tweaked, squashed, and rebased. Closes #1512 ]