Add cloud rings to (some) random Pandemonium Lords
This effect already exists in the game, but only for Asmodeus and Bai
Suzhen (in dragon form). Let random pan lords have a chance at getting
one too.

The chance of a pan lord generating with a cloud ring is about 1 in 12,
which might have to be adjusted upwards or downwards in future. The
chance for each cloud type is a little more complex to calculate given
the dependency on pan lord brand / attack flavour, but the rarest brands
roll about 1 in 12 times.

Depending on the cloud type, these cloud rings can be terrifying or
ignorable. For example, chaotic clouds can instantly paralyse the player
and negative energy clouds can heavily drain them within a few turns. On
the other hand, clouds of fog and mutation do no direct damage.

Pan lords with cloud rings should feel qualitatively different to fight
than those without. Mostly this comes from increased danger being in
melee range, but also from the impact on battlefield terrain that
dropping clouds everywhere provides. In some cases, this can even work
to the benefit of the player.

One aspect of pandemonium lords I hope this change emphasises is that
they should be considered the "apex predator" of DCSS. They should
always be a scary fight, no matter how prepared you are. This is
especially true since they spawn only in either the optional branch of
Pandemonium (where it's even possible to collect the demonic rune
without fighting one) and on the orb run (where running is already
encouraged). This change should increase the number of cases where
players run (in careful ways!) from pan lords.