Nerf condenser vane charges
Player feedback is that the vane recharges too fast and holds too many
charges, so this commit reduces the charges to two and increases the XP
cost of a charge. I went with two charges instead of one because while a
single cloud blast is typically enough in a fight, sometimes a player
might like to reroll their clouds after repositioning, and I want that
option to remain available for the item's use.