Give the Majin-Bo spell vampirism instead of Archmagi
Archmagi for HP was kind of neat, but in practice it was very much a
"get archmagi in not the body slot in exchange for a shield"; it was
rare that you'd be so desparate for the enhancer that the Majin Bo was
your best shot.

Much cooler is to have the Majin Bo make your spells vampiric. All of
the damaging spells. (Note that since Absolute Zero directly kills the
monser bypassing the usual damage code this doesn't interact). The
effect is very powerful, and can be nerfed by reducing the
healing chance or increasing the HP cost if needed.