Fix a seed explorer bug involving portals in branch ends
If a portal is in a branch end, and that branch end immediately precedes
a branch that does not generate, then the seed explorer would repeatedly
regenerate the portal for each following level of a non-generating
branch. This could lead to crashes if the builder ran out of usable
layouts, for example if it tried to build 9 ice caves in a row. It
probably also messed up the seeding for following levels.