Try to break a net holding you when moving up/down stairs
When you are trapped in a net and try to move you instead try to break
the net.  Currently trying to use stairs while you are netted gives "You
can't do that while held in a net."; change this to behave like a
lateral move.

Rationale:  I was netted while on the stairs and I definitely did not
want to step off of them.  I expected "<" to try to break it.  I
actually could have used a lateral move because when you break the net
you don't also do the move, but that isn't obvious.