Adjust and update a hints mode check
This was wildly out of date due to its enum range checks, and also fairly
vague in its definition. The hints mode text specifies that it's for items
that provide "protection from certain sources", so make the list match
that more closely and not return true for other effects that can't really
be described as "resistances" (Faith, Harm etc).