Remove bad position clamp for peeking through stairs
Using X and [ or ] to peek through stairs at different floors is
supposed to deposit the cursor on the other end of the staircase. Lately
the cursor has been frequently deposited in incorrect locations.

It appears that 08ff59a6b inadvertently clamped the destination position
according to the known map bounds of the floor being peeked from, which
tends to move the cursor to an incorrect spot on the floor being peeked
into.

I have simply removed the clamping: if mstate's position was on the same
level then the clamping after the conditional call to goto_level will
serve the same purpose, and if not then clamping before that call is an
error.