Display Rampage on the % screen
It's often hard to see all properties of equipped artefacts on
the overview screen, especially when the artefacts have long names.
The Rampage ego, which can appear on any randart armour, dramatically
changes the player's movement. So it's worth having a quick way
to see if equipped items have this ego.