Trigger untargeted actions directly on autofire
That is, bypass all the convoluted lua stuff when a target is not
needed.

Some of these may do better with an error if there is no foe, but this
new blanket behavior behavior matches the regular cast behavior so it's
a bit closer to what you would expect. Some interesting cases to think
about: blink, berserk (not yet implemented), foxfire, conjure flame,
irradiate. More fine-grained logic needed..